Adds validation for schema_migration prior to truncate.

+++ lib/tasks/gitlab/db.rake
@@ -32,7 +32,7 @@ namespace :gitlab do
       tables = connection.tables
       tables.delete 'schema_migrations'
       # Truncate schema_migrations to ensure migrations re-run
-      connection.execute('TRUNCATE schema_migrations')
+      connection.execute('TRUNCATE schema_migrations') if connection.table_exists? 'schema_migrations' 
 
       # Drop tables with cascade to avoid dependent table errors
       # PG: http://www.postgresql.org/docs/current/static/ddl-depend.html
